requirements.txt
<<<<<<< HEAD
# Core Dependencies
numpy>=1.21.0,<2.0.0
scipy>=1.7.0,<1.14.0
matplotlib>=3.5.0
pandas>=1.3.0

# Simulation and Control
pymavlink>=2.4.37
# airsim>=1.8.1  # Skipped - AirSim not installed in this build
dronekit>=2.9.2
dronekit-sitl>=3.3.0
pybullet>=3.2.5

# Gymnasium (replacement for OpenAI Gym) and RL
gymnasium>=0.29.0
stable-baselines3[extra]>=2.0.0
sb3-contrib>=2.0.0
tensorboard>=2.11.0
shimmy>=0.2.1

# ROS2 Integration (install via rosdep, listed here for reference)
# rclpy
# sensor_msgs
# geometry_msgs
# nav_msgs
# tf2_ros

# Computer Vision
opencv-python>=4.7.0,<4.10.0
Pillow>=9.3.0
scikit-image>=0.19.0

# Path Planning
scikit-learn>=1.0.0
networkx>=2.6.0

# Configuration and Logging
pyyaml>=6.0
python-dotenv>=0.19.0
colorlog>=6.7.0
tqdm>=4.62.0

# Data Processing
h5py>=3.6.0
msgpack>=1.0.3

# Utilities
psutil>=5.9.0
pyserial>=3.5

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Code Quality
black>=22.0.0
flake8>=4.0.0
pylint>=2.12.0
mypy>=0.931

# Visualization
plotly>=5.6.0
seaborn>=0.11.0
dash>=2.3.0

# Performance
numba>=0.55.0
=======
# ~/aerial_photography_drone/requirements.txt に追加
requests>=2.28.0
shapely>=2.0.0
scipy>=1.9.0
>>>>>>> 74e4cbb311d6507a0cb905c6d4851f53cc4aab1d
